requirements.txt
gymnasium==0.29.1
numpy==1.24.3
opencv_python==4.10.0.84
pybullet==3.2.6
stable_baselines3==2.3.2
panda-gym==3.0.7